factored out the tile slicing stuff to be cleaner
