+++ src/entities/HomeBase.ts
@@ -1,45 +1,11 @@
 import { Scene } from 'phaser';
 import {
   HOMEBASE_HP,
+  HOMEBASE_TEXTURE_NAME,
   HOMEBASE__SHAKE_DURATION,
   HOMEBASE__SHAKE_INTENSITY,
-  TILE_SIZE,
 } from '../constants';
 import HealthBar from './HealthBar';
-
-// canvas splicing trick stolen from here:
-// https://codesandbox.io/s/3tuus?file=/src/Play.js
-function sliceFromTexture(
-  scene: Scene,
-  newTextureName: string,
-  sourceTextureName: string,
-  sourceTextureX: number,
-  sourceTextureY: number
-) {
-  const tileSize = TILE_SIZE;
-
-  const grass = scene.textures.get(sourceTextureName);
-  const source = grass.getSourceImage();
-
-  const newT = scene.textures.createCanvas(newTextureName, 32, 32);
-
-  const newCtx = (newT.getSourceImage() as HTMLCanvasElement).getContext('2d');
-
-  // https://developer.mozilla.org/en-US/docs/Web/API/CanvasRenderingContext2D/drawImage
-  newCtx.drawImage(
-    source as HTMLImageElement,
-    sourceTextureX,
-    sourceTextureY,
-    tileSize,
-    tileSize,
-    0,
-    0,
-    tileSize,
-    tileSize
-  );
-
-  newT.refresh();
-}
 
 class HomeBase extends Phaser.GameObjects.Image {
   hp: number;
@@ -49,8 +15,7 @@ class HomeBase extends Phaser.GameObjects.Image {
   originalY: number;
 
   constructor(scene: Scene, x: number, y: number) {
-    sliceFromTexture(scene, 'homebase', 'grass-biome', 32 * 3, 32 * 16);
-    super(scene, 0, 0, 'homebase');
+    super(scene, 0, 0, HOMEBASE_TEXTURE_NAME);
 
     this.x = x;
     this.y = y;
